requirements.txt
discord.py==2.4.0
python-dotenv==1.0.0
asyncio==3.4.3
sqlite3==3.12.5
re
datetime
time
